requirements.txt
azure-core==1.32.0
azure-identity==1.19.0
python-dotenv==1.0.1
pyperclip==1.9.0
# semantic-kernel==1.18.2
semantic-kernel==1.25.0
fastapi
uvicorn
streamlit
openai
requests
pydantic
